pyproject.toml
[tool.poetry]
name = "task_tracker"
version = "0.1.0"
description = "Task tracker helps to assign tasks to employees and track their completion."
authors = ["Natalia Dudina"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.4"
drf-yasg = "^1.21.7"
django = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
